2.8b
Ability to add a server without serverip

+++ AtlasServerManager/Includes/ServerData.cs
@@ -51,35 +51,39 @@ namespace AtlasServerManager.Includes
 
             /* Resolve DNS */
             string CurIP = ServerIp;
-            int DotCount = 0;
-            for (int i = 0; i < CurIP.Length; i++) if (CurIP[i] == '.') DotCount++;
-            if (DotCount != 3)
+            if (CurIP != string.Empty)
             {
-                System.Net.IPAddress[] ips = System.Net.Dns.GetHostAddresses(CurIP);
-                if (ips.Length > 0)
+                int DotCount = 0;
+                for (int i = 0; i < CurIP.Length; i++) if (CurIP[i] == '.') DotCount++;
+                if (DotCount != 3)
                 {
-                    CurIP = ips[0].ToString();
-                    try
+                    System.Net.IPAddress[] ips = System.Net.Dns.GetHostAddresses(CurIP);
+                    if (ips.Length > 0)
                     {
-                        string ServerGrid = Path.Combine(ServerPath + Path.DirectorySeparatorChar, @"ShooterGame\ServerGrid.json");
-                        if (File.Exists(ServerGrid))
+                        CurIP = ips[0].ToString();
+                        try
                         {
-                            string OwnProxGrid = Path.Combine(ServerPath + Path.DirectorySeparatorChar, @"ShooterGame\ServerGridOwnProx.json");
-                            if (File.Exists(OwnProxGrid)) File.Delete(OwnProxGrid);
-                            using (StreamWriter sw = new StreamWriter(OwnProxGrid))
-                            using (StreamReader sr = new StreamReader(ServerGrid))
+                            string ServerGrid = Path.Combine(ServerPath + Path.DirectorySeparatorChar, @"ShooterGame\ServerGrid.json");
+                            if (File.Exists(ServerGrid))
                             {
-                                string line = "";
-                                while ((line = sr.ReadLine()) != null)
+                                string OwnProxGrid = Path.Combine(ServerPath + Path.DirectorySeparatorChar, @"ShooterGame\ServerGridOwnProx.json");
+                                if (File.Exists(OwnProxGrid)) File.Delete(OwnProxGrid);
+                                using (StreamWriter sw = new StreamWriter(OwnProxGrid))
+                                using (StreamReader sr = new StreamReader(ServerGrid))
                                 {
-                                    if (line.Contains("\"ip\": \"")) sw.WriteLine("      \"ip\": \"" + CurIP + "\",");
-                                    else sw.WriteLine(line);
+                                    string line = "";
+                                    while ((line = sr.ReadLine()) != null)
+                                    {
+                                        if (line.Contains("\"ip\": \"")) sw.WriteLine("      \"ip\": \"" + CurIP + "\",");
+                                        else sw.WriteLine(line);
+                                    }
                                 }
+                                File.Delete(ServerGrid);
+                                File.Move(OwnProxGrid, ServerGrid);
                             }
-                            File.Delete(ServerGrid);
-                            File.Move(OwnProxGrid, ServerGrid);
                         }
-                    } catch (Exception e) { MessageBox.Show("Error: " + e.Message, "Error", MessageBoxButtons.OK, MessageBoxIcon.Error); }
+                        catch (Exception e) { MessageBox.Show("Error: " + e.Message, "Error", MessageBoxButtons.OK, MessageBoxIcon.Error); }
+                    }
                 }
             }
 
@@ -92,7 +96,7 @@ namespace AtlasServerManager.Includes
             {
                 ServerProcess = new Process
                 {
-                    StartInfo = new ProcessStartInfo(ExePath, "\"" + "Ocean?ServerX=" + ServerX + "?ServerY=" + ServerY + "?Port=" + ServerPort + "?QueryPort=" + QueryPort + "?AltSaveDirectoryName=" + AltSaveDirectory + "?MaxPlayers=" + MaxPlayers + "?ReservedPlayerSlots=" + ReservedPlayers + "?ServerAdminPassword=" + Pass + "?ServerCrosshair=" + (Crosshair ? "true" : "false") + "?AllowThirdPersonPlayer=" + (Third ? "true" : "false") + "?MapPlayerLocation=" + (MapB ? "true" : "false") + "?serverPVE=" + (!PVP ? "true" : "false") + "?RCONEnabled=" + (Rcon ? ("true?RCONPort=" + RconPort) : "false") + "?EnablePvPGamma=" + (Gamma ? "true" : "false") + "?AllowAnyoneBabyImprintCuddle=" + (Imprint ? "true" : "false") + "?ShowFloatingDamageText=" + FTD + "?SeamlessIP=" + CurIP + CustomArgs + "\" -game -server -log -NoCrashDialog" + (BattleEye ? "" : " -NoBattlEye") + (CustomAfterArgs == string.Empty ? "" : " " + CustomAfterArgs))
+                    StartInfo = new ProcessStartInfo(ExePath, "\"" + "Ocean?ServerX=" + ServerX + "?ServerY=" + ServerY + "?Port=" + ServerPort + "?QueryPort=" + QueryPort + "?AltSaveDirectoryName=" + AltSaveDirectory + "?MaxPlayers=" + MaxPlayers + "?ReservedPlayerSlots=" + ReservedPlayers + "?ServerAdminPassword=" + Pass + "?ServerCrosshair=" + (Crosshair ? "true" : "false") + "?AllowThirdPersonPlayer=" + (Third ? "true" : "false") + "?MapPlayerLocation=" + (MapB ? "true" : "false") + "?serverPVE=" + (!PVP ? "true" : "false") + "?RCONEnabled=" + (Rcon ? ("true?RCONPort=" + RconPort) : "false") + "?EnablePvPGamma=" + (Gamma ? "true" : "false") + "?AllowAnyoneBabyImprintCuddle=" + (Imprint ? "true" : "false") + "?ShowFloatingDamageText=" + FTD + (CurIP == string.Empty ? "" : "?SeamlessIP=" + CurIP) + CustomArgs + "\" -game -server -log -NoCrashDialog" + (BattleEye ? "" : " -NoBattlEye") + (CustomAfterArgs == string.Empty ? "" : " " + CustomAfterArgs))
                     {
                         UseShellExecute = false,
                         WorkingDirectory = Path.GetDirectoryName(ExePath)
